feat: 숫자 리스트 개수, 중복, 범위 검증 메서드 작성 및 사용하는 메서드 작성
docs: 구현한 기능 체크 완료
