+++ docs/README.md
@@ -14,14 +14,14 @@
 - [x] '구입금액을 입력해 주세요.' 출력
 - [ ] 구입금액을 입력 받는다
   - [x] 입력된 금액이 1000원 단위인지 검증하는 기능
-  - [ ] 최소 1000원 ~ 최대 10만원까지 입력받고 그 이하 , 이상은 예외 처리(현실 로또 반영)
+  - [x] 최소 1000원 ~ 최대 10만원까지 입력받고 그 이하 , 이상은 예외 처리(현실 로또 반영)
 - [ ] 사용자가 구입한 개수를 출력한다.
   - [ ] '~개를 구매했습니다.'
 - [ ] 구매한 갯수 만큼 로또 번호를 생성한다.
   - [ ] 지정 범위 내에서 원하는 개수의 수를 무작위로 생성함
     - [x] 생성된 숫자가 지정범위 내인지 검증
-    - [ ] 생성한 숫자 리스트에 중복된 값이 있는지 검증
-    - [ ] 정해진 갯수의 리스트인지 검증
+    - [x] 생성한 숫자 리스트에 중복된 값이 있는지 검증
+    - [x] 정해진 갯수의 리스트인지 검증
     - [ ] 숫자 리스트를 오름차순으로 sort
 - [ ] 랜덤으로 생성된 6개의 로또번호를 사용자가 구매한 갯수 만큼 출력시킨다.
   - [1,2,3,4,5,6] 과 같은 형식으로 출력
